requirements.txt
docling>=1.0.0
click>=8.0.0
pyyaml>=6.0
tqdm>=4.60.0
pathlib2>=2.3.0